Pass http proxy to the server as well
