+++ tsconfig.json
@@ -11,6 +11,7 @@
     "noUnusedParameters": true,
     "noImplicitReturns": true,
     "noFallthroughCasesInSwitch": false,
+    "allowImportingTsExtensions": true,
     "experimentalDecorators": true,
     "inlineSourceMap": true,
     "inlineSources": true,
